fixed a few open issues

+++ ext/ds_game_ui.h
@@ -283,7 +283,7 @@ namespace dialog {
 	// end
 	// ---------------------------------------------------------------
 	void end() {
-		for (size_t i = 0; i < _guiCtx.num_calls; ++i) {
+		for (int i = 0; i < _guiCtx.num_calls; ++i) {
 			const DrawCall& call = _guiCtx.calls[i];
 			_guiCtx.buffer->add(call.pos, call.rect, call.scale);
 		}
